all: update all http.Servers to use a ReadTimeout
Set the `ReadTimeout` on all `http.Server`s in the repository, including
for services:
- Federation
- Friendbot
- Ticker
- Keystore

Why:
It's possible to send a POST request with Content-Length header with a
large value and then send the request body very slowly (ex. 1 byte per
second). When many connections like this are established it's possible
that the server will run out of file descriptors to serve legitimate
users. To solve this we limit the maximum duration for reading the
entire request, including the body to 5 seconds.

Known limitations:
Horizon is not included and is being patched in a separate commit.