services/ticker: reduce size of toml cache in memory (#4289)
Change the maximum size of the toml cache used during asset ingestion in Ticker to a single toml, and sort the assets by their TOML URLs when loading.

If we sort assets by their TOML URLs we only need to store one toml in the cache, the last one. This reduces the amount of memory required.